+++ Tests/TemplatesTests/StoryboardsiOSTests.swift
@@ -37,7 +37,21 @@ class StoryboardsiOSTests: XCTestCase {
        suffix: ""),
       (context: try StencilContext.enrich(context: context,
                                           parameters: ["module=CustomSegue"]),
-       suffix: "-ignore-module")
+       suffix: "-ignore-module"),
+      (context: try StencilContext.enrich(context: context,
+                                          parameters: ["ignoreTargetModule"],
+                                          environment: ["PRODUCT_MODULE_NAME": "Test"]),
+       suffix: ""),
+      (context: try StencilContext.enrich(context: context,
+                                          parameters: ["ignoreTargetModule"],
+                                          environment: ["PRODUCT_MODULE_NAME": "SlackTextViewController"]),
+       suffix: "-ignore-module-need-extra-definitions"),
+      (context: try StencilContext.enrich(context: context,
+                                          parameters: ["module=Test", "ignoreTargetModule"]),
+       suffix: ""),
+      (context: try StencilContext.enrich(context: context,
+                                          parameters: ["module=SlackTextViewController", "ignoreTargetModule"]),
+       suffix: "-ignore-module-need-extra-definitions")
     ]
   }